requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.5
Flask-CORS==4.0.0
python-socketio==5.10.0
python-engineio==4.8.0
humanize==4.9.0
gunicorn==21.2.0
eventlet==0.33.3
google-cloud-bigtable==2.23.0